refactor: [#170] split test_support.rs into focused testing modules
- Split 438-line test_support.rs into 4 focused modules:
  - mod.rs: Module exports and documentation
  - test_writer.rs: TestWriter implementation for shared buffers
  - test_user_output.rs: Core TestUserOutput functionality
  - test_wrapper.rs: TestOutputWrapper convenience methods
- Renamed test_support → testing for consistency with src/testing
- Updated all references across 20+ files in codebase
- Maintained backward compatibility through re-exports
- All 1,178 tests pass with no regressions
- Follows project module organization conventions

Task 1 of issue #170 complete
