+++ src/presentation/controllers/destroy/tests/integration.rs
@@ -10,7 +10,7 @@ use crate::infrastructure::persistence::repository_factory::RepositoryFactory;
 use crate::presentation::controllers::constants::DEFAULT_LOCK_TIMEOUT;
 use crate::presentation::controllers::destroy::{handle_destroy_command, DestroySubcommandError};
 use crate::presentation::controllers::tests::TestContext;
-use crate::presentation::views::test_support::TestUserOutput;
+use crate::presentation::views::testing::TestUserOutput;
 use crate::presentation::views::VerbosityLevel;
 use crate::shared::SystemClock;
